Refactor/ax 982 refactor farm page (#724)
* refactor: used custom styles file for the farm pages and widgets

* refactor: dialogs into separate classes instead of methods

* refactor: convert to super constructors

* refactor: widgets into separate classes instead of methods

* chore: formatting

* refactor: barrel dialogs and widgets exports into one file

* chore: fixing errors

* chore: more formatting
