--- lib/pages/farm/components/no_wallet.dart
@@ -1,14 +1,0 @@
-import 'package:flutter/material.dart';
-
-Widget noWallet() {
-  return const Center(
-    child: SizedBox(
-      height: 70,
-      width: 400,
-      child: Text(
-        'Please connect your wallet.',
-        style: TextStyle(color: Colors.amber, fontSize: 30),
-      ),
-    ),
-  );
-}
