+++ lib/pages/farm/dialogs/dialogs.dart
@@ -1,0 +1,6 @@
+export 'dual_stake_dialog.dart';
+export 'reward_claim_dialog.dart';
+export 'stake_dialog.dart';
+export 'transaction_confirmed_dialog.dart';
+export 'unstake_confirmed_dialog.dart';
+export 'unstake_dialog.dart';
