+++ lib/pages/farm/dialogs/unstake_dialog.dart
@@ -1,8 +1,6 @@
-import 'package:ax_dapp/pages/farm/components/unstake_approve_button.dart';
-import 'package:ax_dapp/pages/farm/dialogs/unstake_confirmed_dialog.dart';
-import 'package:ax_dapp/pages/farm/modules/box_decoration.dart';
-import 'package:ax_dapp/pages/farm/modules/dialog_text_style.dart';
+import 'package:ax_dapp/pages/farm/widgets/widgets.dart';
 import 'package:ax_dapp/service/controller/farms/farm_controller.dart';
+import 'package:ax_dapp/service/custom_styles.dart';
 import 'package:ax_dapp/util/warning_text_button.dart';
 import 'package:ax_dapp/wallet/bloc/wallet_bloc.dart';
 import 'package:flutter/material.dart';
@@ -71,7 +69,12 @@ class _UnstakeDialogState extends State<UnstakeDialog> {
               children: [
                 Text(
                   'Unstake Liquidity',
-                  style: textStyle(Colors.white, 20, false),
+                  style: textStyle(
+                    Colors.white,
+                    20,
+                    isBold: false,
+                    isUline: false,
+                  ),
                 ),
                 TextButton(
                   onPressed: () {
@@ -120,7 +123,12 @@ class _UnstakeDialogState extends State<UnstakeDialog> {
                       Expanded(
                         child: Text(
                           '''${selectedFarm.strStakedAlias.value.isNotEmpty ? selectedFarm.strStakedAlias : selectedFarm.strStakedSymbol}''',
-                          style: textStyle(Colors.white, 15, false),
+                          style: textStyle(
+                            Colors.white,
+                            15,
+                            isBold: false,
+                            isUline: false,
+                          ),
                         ),
                       ),
                       Container(
@@ -141,7 +149,12 @@ class _UnstakeDialogState extends State<UnstakeDialog> {
                           },
                           child: Text(
                             'Max',
-                            style: textStyle(Colors.grey[400]!, 9, false),
+                            style: textStyle(
+                              Colors.grey[400]!,
+                              9,
+                              isBold: false,
+                              isUline: false,
+                            ),
                           ),
                         ),
                       ),
@@ -159,10 +172,20 @@ class _UnstakeDialogState extends State<UnstakeDialog> {
                               selectedFarm,
                             );
                           },
-                          style: textStyle(Colors.grey[400]!, 22, false),
+                          style: textStyle(
+                            Colors.grey[400]!,
+                            22,
+                            isBold: false,
+                            isUline: false,
+                          ),
                           decoration: InputDecoration(
                             hintText: '0.00',
-                            hintStyle: textStyle(Colors.grey[400]!, 22, false),
+                            hintStyle: textStyle(
+                              Colors.grey[400]!,
+                              22,
+                              isBold: false,
+                              isUline: false,
+                            ),
                             contentPadding: const EdgeInsets.all(9),
                             border: InputBorder.none,
                           ),
@@ -183,12 +206,22 @@ class _UnstakeDialogState extends State<UnstakeDialog> {
               children: [
                 Text(
                   'Current ${selectedFarm.strStakedSymbol} Staked',
-                  style: textStyle(Colors.grey[400]!, 14, false),
+                  style: textStyle(
+                    Colors.grey[400]!,
+                    14,
+                    isBold: false,
+                    isUline: false,
+                  ),
                 ),
                 Obx(
                   () => Text(
                     '''${selectedFarm.stakedInfo.value.viewAmount} ${selectedFarm.strStakedSymbol}''',
-                    style: textStyle(Colors.grey[400]!, 14, false),
+                    style: textStyle(
+                      Colors.grey[400]!,
+                      14,
+                      isBold: false,
+                      isUline: false,
+                    ),
                   ),
                 ),
               ],
@@ -199,7 +232,12 @@ class _UnstakeDialogState extends State<UnstakeDialog> {
                   padding: const EdgeInsets.only(left: 55),
                   child: Text(
                     '-',
-                    style: textStyle(Colors.grey[400]!, 14, false),
+                    style: textStyle(
+                      Colors.grey[400]!,
+                      14,
+                      isBold: false,
+                      isUline: false,
+                    ),
                   ),
                 )
               ],
@@ -209,12 +247,22 @@ class _UnstakeDialogState extends State<UnstakeDialog> {
               children: [
                 Text(
                   'Funds Removed',
-                  style: textStyle(Colors.grey[400]!, 14, false),
+                  style: textStyle(
+                    Colors.grey[400]!,
+                    14,
+                    isBold: false,
+                    isUline: false,
+                  ),
                 ),
                 Obx(
                   () => Text(
                     '''${selectedFarm.strUnStakeInput.value} ${selectedFarm.strStakedSymbol}''',
-                    style: textStyle(Colors.grey[400]!, 14, false),
+                    style: textStyle(
+                      Colors.grey[400]!,
+                      14,
+                      isBold: false,
+                      isUline: false,
+                    ),
                   ),
                 ),
               ],
@@ -246,7 +294,6 @@ class _UnstakeDialogState extends State<UnstakeDialog> {
                       width: 175,
                       height: 45,
                       text: 'Confirm',
-                      confirmDialog: unstakeConfirmedDialog,
                       selectedFarm: selectedFarm,
                       walletAddress: context
                           .read<WalletBloc>()
